Removed unecessary/unused file fastapi_app.py Refactored queries in rag_pipeline.py to prompts.py to improve readability.
